Pushing Sammy 0.4.1

+++ Rakefile
@@ -76,7 +76,7 @@ task :api do
   tmp_doc_path = '/tmp/sammy.api.html'
   api_template_path = 'site/docs/api_template.html'
   final_path   = 'site/docs/api.html'
-  File.unlink(tmp_doc_path)
+  File.unlink(tmp_doc_path) if File.readable?(tmp_doc_path)
   sh "ruby vendor/jsdoc/jsdoc.rb lib/sammy.js lib/plugins/ > #{tmp_doc_path}"
   sh "cat #{api_template_path} #{tmp_doc_path} > #{final_path}"
 end
